boost 1.83

+++ src/cxx_supportlib/vendor-modified/boost/throw_exception.hpp
@@ -104,9 +104,9 @@ public:
         copy_from( &e );
 
         set_info( *this, throw_file( loc.file_name() ) );
-        set_info( *this, throw_line( loc.line() ) );
+        set_info( *this, throw_line( static_cast<int>( loc.line() ) ) );
         set_info( *this, throw_function( loc.function_name() ) );
-        set_info( *this, throw_column( loc.column() ) );
+        set_info( *this, throw_column( static_cast<int>( loc.column() ) ) );
     }
 
     virtual boost::exception_detail::clone_base const * clone() const BOOST_OVERRIDE
